Using gl_utils

+++ gl_utils.hpp
@@ -3,12 +3,12 @@
 #include "matrix4.hpp"
 
 namespace MyGL {
-    Matrix4 lookAt(const double&, const double&,
-                const double&, const double&,
-                const double&, const double&,
-                const double&, const double&,
-                const double&);
+    Matrix4 lookAt(const float&, const float&,
+                   const float&, const float&,
+                   const float&, const float&,
+                   const float&, const float&,
+                   const float&);
 
-    Matrix4 frustum(double, double, double,
-                   double, double, double);
+    Matrix4 frustum(float, float, float,
+                    float, float, float);
 }
